Cleanup, update trend analysis in visualization
