+++ web/src/pages/ProductPage.scss
@@ -26,6 +26,6 @@ span.price {
     filter: brightness(1.5);
 }
 
-span.price, span.average, span.historical-average, span.threshold, span.long-slope {
+span.price, span.average, span.thirty-day-average, span.threshold, span.seven-day-trend {
     margin-right: 24px;
 }
